agents/requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-anthropic>=0.2.0
anthropic>=0.34.0
mcp>=1.12.0
pydantic>=2.8.0
python-dotenv>=1.0.0
asyncio-mqtt==0.16.1
websockets>=12.0
httpx>=0.25.0
fastapi>=0.100.0
uvicorn>=0.20.0 